// types/enums/onec.ts
/**
 * 1C:Enterprise metadata types and enums
 */

export enum OnecDirName {
  Catalogs = 'Catalogs',
  Documents = 'Documents',
  Enums = 'Enums',
  Reports = 'Reports',
  CommonModules = 'CommonModules',
  InformationRegisters = 'InformationRegisters',
  AccumulationRegisters = 'AccumulationRegisters',
  CalculationRegisters = 'CalculationRegisters',
  BusinessProcesses = 'BusinessProcesses',
  Tasks = 'Tasks',
  Roles = 'Roles',
  Subsystems = 'Subsystems',
  Constants = 'Constants',
  DataProcessors = 'DataProcessors',
  CommonForms = 'CommonForms',
  HTTPServices = 'HTTPServices',
  WebServices = 'WebServices',
  WSReferences = 'WSReferences',
  Languages = 'Languages',
  ChartsOfCalculationTypes = 'ChartsOfCalculationTypes',
  ChartsOfCharacteristicTypes = 'ChartsOfCharacteristicTypes',
  DocumentJournals = 'DocumentJournals',
  DocumentNumerators = 'DocumentNumerators',
  ExchangePlans = 'ExchangePlans',
  DefinedTypes = 'DefinedTypes',
  FilterCriteria = 'FilterCriteria',
  EventSubscriptions = 'EventSubscriptions',
  FunctionalOptions = 'FunctionalOptions',
  FunctionalOptionsParameters = 'FunctionalOptionsParameters',
  CommonCommands = 'CommonCommands',
  CommonAttributes = 'CommonAttributes',
  SettingsStorages = 'SettingsStorages',
  StyleItems = 'StyleItems',
  CommandGroups = 'CommandGroups',
  ScheduledJobs = 'ScheduledJobs',
  SessionParameters = 'SessionParameters',
  XDTOPackages = 'XDTOPackages',
}

export enum OnecNodeType {
  Catalog = 'Catalog',
  Document = 'Document',
  Enum = 'Enum',
  Report = 'Report',
  CommonModule = 'CommonModule',
  InformationRegister = 'InformationRegister',
  AccumulationRegister = 'AccumulationRegister',
  CalculationRegister = 'CalculationRegister',
  BusinessProcess = 'BusinessProcess',
  Task = 'Task',
  Role = 'Role',
  Subsystem = 'Subsystem',
  Constant = 'Constant',
  DataProcessor = 'DataProcessor',
  CommonForm = 'CommonForm',
  HTTPService = 'HTTPService',
  WebService = 'WebService',
  WSReference = 'WSReference',
  Language = 'Language',
  ChartOfCalculationTypes = 'ChartOfCalculationTypes',
  ChartOfCharacteristicTypes = 'ChartOfCharacteristicTypes',
  DocumentJournal = 'DocumentJournal',
  DocumentNumerator = 'DocumentNumerator',
  ExchangePlan = 'ExchangePlan',
  DefinedType = 'DefinedType',
  FilterCriteria = 'FilterCriteria',
  EventSubscription = 'EventSubscription',
  FunctionalOption = 'FunctionalOption',
  FunctionalOptionParameter = 'FunctionalOptionParameter',
  CommonCommand = 'CommonCommand',
  CommonAttribute = 'CommonAttribute',
  SettingsStorage = 'SettingsStorage',
  StyleItem = 'StyleItem',
  CommandGroup = 'CommandGroup',
  ScheduledJob = 'ScheduledJob',
  SessionParameter = 'SessionParameter',
  XDTOPackage = 'XDTOPackage',
}